+++ examples/get-group-schedule.ts
@@ -6,11 +6,7 @@ const groupNumber = process.argv[2] ?? "053503";
 
 async function main(): Promise<void> {
   const schedule = await client.schedule.getGroup(groupNumber);
-  if ("lessons" in schedule) {
-    console.log(`Lessons count for ${groupNumber}:`, schedule.lessons.length);
-  } else {
-    console.log(`Exams count for ${groupNumber}:`, schedule.exams.length);
-  }
+  console.log(`Lessons count for ${groupNumber}:`, schedule.lessons.length);
 }
 
 void main();